test uav matching

+++ Block_wise_GPS_correction.py
@@ -2541,7 +2541,7 @@ class Patch:
 
 		img = green_channel-0.61*blue_channel-0.39*red_channel
 		img[img<0] = 0
-		img = img.astype('uint8')
+		# img = img.astype('uint8')
 
 		min_p = np.amin(img)
 		max_p = np.amax(img)
